Add JQ_REPL_ARGS environment variable
This is so that we can pass JQ flags to the REPL. Sometimes it is
necessary to use jq in ways that can only be influenced by flags. Things
like slurping (-s), printing raw strings (-r), parsing raw input (-R)
and various other flags that affect input parsing or output printing.